M6: Rule-based optimizer and type normalisation
Optimizer framework (lumina-plan/optimizer):
- Rule trait: single rewrite applied bottom-up by Optimizer
- Optimizer.optimize(plan, rules): applies rules after recursing into
  children; defaultRules = [CombineFilters, PredicatePushdown]
- CombineFilters: Filter(Filter(child, c2), c1) → Filter(child, And(c1,c2))
  gives DuckDB a single WHERE clause for cross-predicate optimisation
- PredicatePushdown: moves Filter below Project (when all referenced
  columns are projected) and below Sort (always safe); Aggregate and
  ReadCsv boundaries are left in place

Type normalisation (lumina-plan/backend/RowNormalizer):
- Normalises JDBC boxed types (java.lang.Integer, Long, Double, Float,
  Boolean, Short, Byte, BigDecimal) to Scala primitives after DuckDB
  result collection; callers can now write backend-independent assertions
- DuckDBBackend applies RowNormalizer.normalizeAll on every result set

Wiring:
- LocalBackend and DuckDBBackend both call Optimizer.optimize(plan)
  before execution; the optimized plan is what gets interpreted / translated

Tests (179 total, 0 failures):
- CombineFiltersSpec: 4 tests for merge and no-op cases
- PredicatePushdownSpec: 7 tests covering pushthrough Project/Sort,
  blocking at Aggregate, and the combined optimizer pipeline
- RowNormalizerSpec: 11 tests, one per Java type + passthrough cases

+++ lumina-backend-duckdb/src/test/scala/lumina/backend/duckdb/DuckDBBackendM5Spec.scala
@@ -117,7 +117,8 @@ class DuckDBBackendM5Spec extends FunSuite:
       schema       = None
     )
     val berlin = rows(plan).find(_.values("city") == "Berlin").get
-    assertEquals(berlin.values("min_age").asInstanceOf[Integer].intValue(), 22)
+    // RowNormalizer converts JDBC Integer to Scala Int
+    assertEquals(berlin.values("min_age").asInstanceOf[Int], 22)
 
   test("Max aggregation executed via DuckDB returns the maximum value per group"):
     val plan = Aggregate(
